extract fns to serialize results to db to worker.serializers
