+++ depobs/website/config.py
@@ -23,6 +23,7 @@ LOGGING = {
         "do": {"handlers": ["console"], "level": "DEBUG"},
         "request.summary": {"handlers": ["console"], "level": "INFO"},
         "depobs.database.models": {"handlers": ["console"], "level": "INFO"},
+        "depobs.database.serializers": {"handlers": ["console"], "level": "INFO"},
         "depobs.website.views": {"handlers": ["console"], "level": "INFO"},
         "depobs.website.scans": {"handlers": ["console"], "level": "INFO"},
         "depobs.website.score_details.blueprint": {
@@ -45,10 +46,6 @@ LOGGING = {
             "level": "INFO",
         },
         "depobs.scanner.pipelines.run_repo_tasks": {
-            "handlers": ["console"],
-            "level": "INFO",
-        },
-        "depobs.scanner.pipelines.postprocess": {
             "handlers": ["console"],
             "level": "INFO",
         },
