+++ setup.cfg
@@ -31,7 +31,6 @@ files =
       depobs/scanner/pipelines/find_dep_files.py,
       depobs/scanner/pipelines/find_git_refs.py,
       depobs/scanner/pipelines/github_metadata.py,
-      depobs/scanner/pipelines/postprocess.py,
       depobs/scanner/pipelines/run_repo_tasks.py,
       depobs/scanner/pipelines/rust_changelog.py,
       depobs/scanner/pipelines/util.py,
